+++ packages/unified-latex-util-macros/libs/newcommand.ts
@@ -4,6 +4,7 @@ import { printRaw } from "@unified-latex/unified-latex-util-print-raw";
 import { replaceNode } from "@unified-latex/unified-latex-util-replace";
 import { visit } from "@unified-latex/unified-latex-util-visit";
 import { getNamedArgsContent } from "@unified-latex/unified-latex-util-arguments";
+import { parseMinimal } from "@unified-latex/unified-latex-util-parse";
 import {
     HashNumber,
     parseMacroSubstitutions,
@@ -89,7 +90,7 @@ export function newcommandMacroToSpec(node: Ast.Macro): string {
         // If it is present, we need to change the signature.
         if (namedArgs.default != null) {
             numArgsForSig--;
-            sigOptionalArg = ["o"];
+            sigOptionalArg = [`O{${printRaw(namedArgs.default)}}`];
         }
         return [
             ...sigOptionalArg,
@@ -181,6 +182,25 @@ export function newcommandMacroToSubstitutionAst(node: Ast.Macro): Ast.Node[] {
     return [];
 }
 
+type ExpanderArg = {
+    content: Ast.Node[];
+    /**
+     * The hash numbers that are used in this argument.
+     */
+    hashNumbers: number[];
+};
+
+/**
+ * A macro can have arguments `#1`...`#9`. This function returns an array of
+ * `Ast.Node` that can be used as the default arguments for a macro if none are provided.
+ */
+function defaultExpanderArgs(): ExpanderArg[] {
+    return Array.from({ length: 10 }).map((_, i) => ({
+        hashNumbers: [],
+        content: [{ type: "string", content: `#${i + 1}` }],
+    }));
+}
+
 /**
  * A factory function. Given a macro definition, creates a function that accepts
  * the macro's arguments and outputs an Ast with the contents substituted (i.e.,
@@ -190,15 +210,96 @@ export function createMacroExpander(
     substitution: Ast.Node[]
 ): (macro: Ast.Macro) => Ast.Node[] {
     const cachedSubstitutionTree = structuredClone(substitution);
-    let hasSubstitutions = false;
+    const hashNumbers = parseHashNumbers(cachedSubstitutionTree);
+
+    return (macro: Ast.Macro) => {
+        // We can short-circuit if there are no substitutions to do.
+        if (hashNumbers.length === 0) {
+            return structuredClone(cachedSubstitutionTree);
+        }
+
+        const cachedSubstitutions: ExpanderArg[] = defaultExpanderArgs().map(
+            (expanderArg, i) => {
+                const number = i + 1;
+                if (!hashNumbers.includes(number)) {
+                    // If we're here, the hash number `i+1` has not been requested for use in a substitution.
+                    // It is safe to return the default (it will never be used)
+                    return expanderArg;
+                }
+                const arg = macro.args?.[i];
+                const defaultArg = arg?._renderInfo?.defaultArg;
+                if (!arg || (isEmptyArg(arg) && defaultArg != null)) {
+                    const content = cachedParse(defaultArg!);
+                    const hashNumbers = parseHashNumbers(content);
+                    return {
+                        content,
+                        hashNumbers,
+                    };
+                }
+                return { content: arg.content, hashNumbers: [] };
+            }
+        );
+
+        // Since we are only allowed `#1`...`#9`, it can take at most 10 iterations to expand in the worst case.
+        // If we expand further than that, there is a circular reference.
+        let numTimesExpanded = 0;
+        while (
+            expandCachedSubstitutions(cachedSubstitutions) &&
+            numTimesExpanded < cachedSubstitutions.length
+        ) {
+            numTimesExpanded++;
+        }
+
+        // By this point we have expanded substitutions as much as possible.
+        // If there are any remaining hash numbers, they come from circular references.
+        for (const expanderArg of cachedSubstitutions) {
+            if (expanderArg.hashNumbers.length > 0) {
+                expanderArg.content = [
+                    // `xparse` seems to use `-No Value-` here.
+                    { type: "string", content: `-Circular-` },
+                ];
+            }
+        }
+
+        const retTree = structuredClone(cachedSubstitutionTree);
+        replaceNode(retTree, (node) => {
+            const hashNumOrNode = node as Ast.Node | HashNumber;
+            if (hashNumOrNode.type !== "hash_number") {
+                return;
+            }
+            return cachedSubstitutions[hashNumOrNode.number - 1].content;
+        });
+        return retTree;
+    };
+}
+
+/**
+ * Is the argument empty? This occurs when optional arguments are not provided.
+ */
+function isEmptyArg(arg: Ast.Argument) {
+    return (
+        arg.openMark === "" && arg.closeMark === "" && arg.content.length === 0
+    );
+}
+
+/**
+ * Parses `tree` for hash numbers (e.g. `#1` or `##`, etc.). Hash numbers are replaced
+ * with a `{type: "hash_number"}` node. This mutates `tree`.
+ * @param tree
+ * @returns A list containing all hash numbers that were found.
+ */
+function parseHashNumbers(tree: Ast.Node[]) {
+    let hashNumbers = new Set<number>();
     visit(
-        cachedSubstitutionTree,
+        tree,
         (nodes) => {
             const parsed = parseMacroSubstitutions(nodes);
-            // Keep track of whether there are any substitutions so we can bail early if not.
-            hasSubstitutions =
-                hasSubstitutions ||
-                parsed.some((node) => node.type === "hash_number");
+            for (const node of parsed) {
+                if (node.type === "hash_number") {
+                    hashNumbers.add(node.number);
+                }
+            }
+            // Replace the nodes with the parsed version
             nodes.length = 0;
             nodes.push(...(parsed as Ast.Node[]));
         },
@@ -207,30 +308,55 @@ export function createMacroExpander(
             test: Array.isArray,
         }
     );
+    return Array.from(hashNumbers);
+}
 
-    return (macro: Ast.Macro) => {
-        if (!hasSubstitutions) {
-            return cachedSubstitutionTree;
+/**
+ * Get a list of all hash numbers referenced within `tree`.
+ */
+function hashNumbersReferenced(tree: Ast.Node[]) {
+    let hashNumbers = new Set<number>();
+    visit(tree, (node) => {
+        const n = node as any as Ast.Node | HashNumber;
+        if (n.type === "hash_number") {
+            hashNumbers.add(n.number);
         }
-        const cachedSubstitutions = (macro.args || []).map(
-            (arg) => arg.content
-        );
-        function getSubstitutionForHashNumber(hashNumber: HashNumber) {
-            return (
-                cachedSubstitutions[hashNumber.number - 1] || {
-                    type: "string",
-                    content: `#${hashNumber.number}`,
-                }
-            );
+    });
+    return Array.from(hashNumbers);
+}
+
+const parseCache = new Map<string, Ast.Node[]>();
+/**
+ * Parse `source` and cache the result. Multiple on the same string will not reparse.
+ */
+function cachedParse(source: string): Ast.Node[] {
+    const cached = parseCache.get(source);
+    if (cached) {
+        return structuredClone(cached);
+    }
+    const parsed = parseMinimal(source).content;
+    parseCache.set(source, structuredClone(parsed));
+    return parsed;
+}
+
+/**
+ * Perform one step in the expansion of `expanderArgs`. The result may still have `type: "hash_number"` nodes.
+ */
+function expandCachedSubstitutions(expanderArgs: ExpanderArg[]) {
+    let didExpand = false;
+    for (const expanderArg of expanderArgs) {
+        if (expanderArg.hashNumbers.length === 0) {
+            continue;
         }
-        const retTree = structuredClone(cachedSubstitutionTree);
-        replaceNode(retTree, (node) => {
+        replaceNode(expanderArg.content, (node) => {
             const hashNumOrNode = node as Ast.Node | HashNumber;
             if (hashNumOrNode.type !== "hash_number") {
                 return;
             }
-            return getSubstitutionForHashNumber(hashNumOrNode);
+            didExpand = true;
+            return expanderArgs[hashNumOrNode.number - 1].content;
         });
-        return retTree;
-    };
+        expanderArg.hashNumbers = hashNumbersReferenced(expanderArg.content);
+    }
+    return didExpand;
 }